Impl ProductService

+++ src/main/java/entity/Boots.java
@@ -4,7 +4,7 @@ public class Boots extends Product {
     private int size;
     private boolean isNaturalSkin;
 
-    public Boots(int id, String productName, int price, int weight, String color, int productCount,
+    public Boots(Long id, String productName, int price, int weight, String color, int productCount,
                  int size, boolean isNaturalSkin){
         super(id, productName, price, weight, color, productCount);
         this.size = size;
